Add files via upload
This julia file contains functions that: initialize a reaction network utilizing catalyst, generate a normal distribution of each parameter of the reaction network, visualize these distributions, and run the simulation for a user defined period of time, save points, and number of loops. The outputs of the run_simulations function are the time traces of each of the simulated species and the parameters that were randomly selected for that iteration of the loop used by the simulation. It also includes some code at the end that can be used to quickly visualize the output of the simulations.